.github/actions/check-holiday-updates/requirements.txt
PyGithub==2.8.1
python-dateutil==2.9.0
